fix: market depth from order event changes
