+++ cmd/data-node/commands/start/sqlsubscribers.go
@@ -122,9 +122,9 @@ type SQLSubscribers struct {
 	keyRotationsSub         *sqlsubscribers.KeyRotation
 	ethereumKeyRotationsSub *sqlsubscribers.EthereumKeyRotation
 	nodeSub                 *sqlsubscribers.Node
-	marketDepthSub          *sqlsubscribers.MarketDepth
-	pupSub                  *sqlsubscribers.ProtocolUpgrade
-	snapSub                 *sqlsubscribers.SnapshotData
+	// marketDepthSub          *sqlsubscribers.MarketDepth
+	pupSub  *sqlsubscribers.ProtocolUpgrade
+	snapSub *sqlsubscribers.SnapshotData
 }
 
 func (s *SQLSubscribers) GetSQLSubscribers() []broker.SQLBrokerSubscriber {
@@ -162,7 +162,7 @@ func (s *SQLSubscribers) GetSQLSubscribers() []broker.SQLBrokerSubscriber {
 		s.multiSigSignerEventSub,
 		s.keyRotationsSub,
 		s.nodeSub,
-		s.marketDepthSub,
+		// s.marketDepthSub,
 		s.ethereumKeyRotationsSub,
 		s.pupSub,
 		s.snapSub,
@@ -270,7 +270,7 @@ func (s *SQLSubscribers) SetupSQLSubscribers() {
 	s.assetSub = sqlsubscribers.NewAsset(s.assetService)
 	s.partySub = sqlsubscribers.NewParty(s.partyService)
 	s.transferResponseSub = sqlsubscribers.NewTransferResponse(s.ledgerService, s.accountService)
-	s.orderSub = sqlsubscribers.NewOrder(s.orderService)
+	s.orderSub = sqlsubscribers.NewOrder(s.orderService, s.marketDepthService)
 	s.networkLimitsSub = sqlsubscribers.NewNetworkLimitSub(s.networkLimitsService)
 	s.marketDataSub = sqlsubscribers.NewMarketData(s.marketDataService)
 	s.tradesSub = sqlsubscribers.NewTradesSubscriber(s.tradeService)
@@ -298,7 +298,7 @@ func (s *SQLSubscribers) SetupSQLSubscribers() {
 	s.keyRotationsSub = sqlsubscribers.NewKeyRotation(s.keyRotationsService)
 	s.ethereumKeyRotationsSub = sqlsubscribers.NewEthereumKeyRotation(s.ethereumKeyRotationsService)
 	s.nodeSub = sqlsubscribers.NewNode(s.nodeService)
-	s.marketDepthSub = sqlsubscribers.NewMarketDepth(s.marketDepthService)
+	// s.marketDepthSub = sqlsubscribers.NewMarketDepth(s.marketDepthService)
 	s.pupSub = sqlsubscribers.NewProtocolUpgrade(s.protocolUpgradeService)
 	s.snapSub = sqlsubscribers.NewSnapshotData(s.coreSnapshotService)
 }
